+++ packages/mapbox/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# v1.0.14 - March 1, 2021
+
+- add `getFeatureGeometryType`, `getViewportForFeature`, `getViewportForBounds` to exports.
+- fix bug where default viewport was not set.
+
 # v1.0.12 - November 19, 2020
 
 - added `maxBounds` prop to specify map bounds